Fix upgrade test

+++ test/setup_test.go
@@ -431,10 +431,16 @@ func applyAndWaitForApplications(commitID string) {
 	// TODO: remove this block after release the PR bellow
 	// https://github.com/cybozu-go/neco-apps/pull/2123
 	if doUpgrade {
-		stdout, stderr, err := ExecAt(boot0, "kubectl", "annotate", "cephblockpool", "-n", "ceph-hdd", "ceph-hdd-block-pool", "admission.cybozu.com/i-am-sure-to-delete=ceph-hdd-block-pool")
-		Expect(err).ShouldNot(HaveOccurred(), "failed to annotate: stdout=%s, stderr=%s", stdout, stderr)
-		stdout, stderr, err = ExecAt(boot0, "kubectl", "annotate", "cephobjectstore", "-n", "ceph-hdd", "ceph-hdd-object-store", "admission.cybozu.com/i-am-sure-to-delete=ceph-hdd-object-store")
-		Expect(err).ShouldNot(HaveOccurred(), "failed to annotate: stdout=%s, stderr=%s", stdout, stderr)
+		_, _, err := ExecAt(boot0, "kubectl", "get", "cephblockpool", "-n", "ceph-hdd", "ceph-hdd-block-pool")
+		if err == nil {
+			stdout, stderr, err := ExecAt(boot0, "kubectl", "annotate", "cephblockpool", "-n", "ceph-hdd", "ceph-hdd-block-pool", "admission.cybozu.com/i-am-sure-to-delete=ceph-hdd-block-pool")
+			Expect(err).ShouldNot(HaveOccurred(), "failed to annotate: stdout=%s, stderr=%s", stdout, stderr)
+		}
+		_, _, err = ExecAt(boot0, "kubectl", "get", "cephobjectstore", "-n", "ceph-hdd", "ceph-hdd-object-store")
+		if err == nil {
+			stdout, stderr, err := ExecAt(boot0, "kubectl", "annotate", "cephobjectstore", "-n", "ceph-hdd", "ceph-hdd-object-store", "admission.cybozu.com/i-am-sure-to-delete=ceph-hdd-object-store")
+			Expect(err).ShouldNot(HaveOccurred(), "failed to annotate: stdout=%s, stderr=%s", stdout, stderr)
+		}
 	}
 
 	By("syncing argocd-config")
